scripts/url-audit-tool/requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas>=1.3.0
PyYAML>=5.4.1
tqdm>=4.62.0
urllib3>=1.26.5
